+++ marxs/source/tests/test_labPointSource.py
@@ -10,7 +10,7 @@ from ..labSource import LabPointSourceCone
 def test_labsource_eukledian():
     with pytest.raises(ValueError) as err:
         source = LabPointSourceCone(position=[1, 2, 3, 1.])
-    assert 'in Eukledian coordinates' in str(err.value)
+    assert "in Euclidean coordinates" in str(err.value)
 
 
 def test_photon_generation():
@@ -26,10 +26,10 @@ def test_photon_generation():
 
 
 def test_directions_range_cone():
-    '''This tests that all of the photons in the returned table have directions within the given parameters.
-    - Ensures that all direciton vectors are normed
+    """This tests that all of the photons in the returned table have directions within the given parameters.
+    - Ensures that all direction vectors are normed
     - Ensures that all direction vectors are within specified range
-    '''
+    """
 
     # parameters
     pos = [5 * np.random.random(), 5 * np.random.random(), 5 * np.random.random()]